+++ packaging/cmake/try_compile/rdkafka_setup.cmake
@@ -32,6 +32,13 @@ try_compile(
     LINK_LIBRARIES "-lpthread"
 )
 
+try_compile(
+    HAVE_PTHREAD_SETNAME_FREEBSD
+    "${CMAKE_CURRENT_BINARY_DIR}/try_compile"
+    "${TRYCOMPILE_SRC_DIR}/pthread_setname_freebsd_test.c"
+    LINK_LIBRARIES "-lpthread"
+)
+
 # Atomic 32 tests {
 set(LINK_ATOMIC NO)
 set(HAVE_ATOMICS_32 NO)